Move linkUser
