+++ realm/realm-library/src/androidTestObjectServer/kotlin/io/realm/RealmAppTests.kt
@@ -275,90 +275,6 @@ class RealmAppTests {
         }
     }
 
-    @Ignore("FIXME: Wait for linkUser support in ObjectStore")
-    @Test
-    fun linkUser() {
-        admin.setAutomaticConfirmation(enabled = false)
-        val user: RealmUser = app.login(RealmCredentials.anonymous())
-        assertEquals(1, user.identities.size)
-        val email = TestHelper.getRandomEmail()
-        val password = "123456"
-        app.emailPasswordAuthProvider.registerUser(email, password) // TODO: Test what happens if auto-confirm is enabled
-        val linkedUser: RealmUser = app.linkUser(RealmCredentials.emailPassword(email, password))
-        assertTrue(user === linkedUser)
-        assertEquals(2, linkedUser.identities.size)
-        assertEquals(RealmCredentials.IdentityProvider.EMAIL_PASSWORD, linkedUser.identities[1].provider)
-        admin.setAutomaticConfirmation(enabled = true)
-    }
-
-    @Ignore("FIXME: Wait for linkUser support in ObjectStore")
-    @Test
-    fun linkUser_existingCredentialsThrows() {
-        admin.setAutomaticConfirmation(enabled = false)
-        val email = TestHelper.getRandomEmail()
-        val password = "123456"
-        val emailUser: RealmUser = app.registerUserAndLogin(email, password)
-        val anonymousUser: RealmUser = app.login(RealmCredentials.anonymous())
-        try {
-            app.linkUser(RealmCredentials.emailPassword(email, password))
-            fail()
-        } catch (ex: ObjectServerError) {
-            assertEquals(ErrorCode.BAD_REQUEST, ex.errorCode)
-        }
-    }
-
-    @Ignore("FIXME: Wait for linkUser support in ObjectStore")
-    @Test
-    fun linkUser_noCurrentUserThrows() {
-        try {
-            app.linkUser(RealmCredentials.emailPassword(TestHelper.getRandomEmail(), "123456"))
-            fail()
-        } catch (ignore: IllegalStateException) {
-        }
-    }
-
-    @Ignore("FIXME: Wait for linkUser support in ObjectStore")
-    @Test
-    fun linkUser_invalidArgsThrows() {
-        try {
-            app.linkUser(TestHelper.getNull())
-            fail()
-        } catch (ignore: IllegalArgumentException) {
-        }
-    }
-
-    @Ignore("FIXME: Wait for linkUser support in ObjectStore")
-    @Test
-    fun linkUserAsync() {
-        admin.setAutomaticConfirmation(enabled = false)
-        val user: RealmUser = app.login(RealmCredentials.anonymous())
-        assertEquals(1, user.identities.size)
-        val email = TestHelper.getRandomEmail()
-        val password = "123456"
-        app.emailPasswordAuthProvider.registerUser(email, password) // TODO: Test what happens if auto-confirm is enabled
-        looperThread.runBlocking {
-            app.linkUserAsync(RealmCredentials.emailPassword(email, password)) { result ->
-                val linkedUser: RealmUser = result.orThrow
-                assertTrue(user === linkedUser)
-                assertEquals(2, linkedUser.identities.size)
-                assertEquals(RealmCredentials.IdentityProvider.EMAIL_PASSWORD, linkedUser.identities[1].provider)
-                admin.setAutomaticConfirmation(enabled = true)
-            }
-        }
-    }
-
-    @Ignore("FIXME: Wait for linkUser support in ObjectStore")
-    @Test
-    fun linkUserAsync_throwsOnNonLooperThread() {
-        val user: RealmUser = app.login(RealmCredentials.anonymous())
-        try {
-            app.linkUserAsync(RealmCredentials.emailPassword(TestHelper.getRandomEmail(), "123456")) { fail() }
-            fail()
-        } catch (ignore: java.lang.IllegalStateException) {
-        }
-
-    }
-
     @Test
     fun getApiKeyAuthProvider() {
         val user1: RealmUser = app.registerUserAndLogin(TestHelper.getRandomEmail(), "123456")
